+++ awint_internals/src/lib.rs
@@ -33,12 +33,15 @@ pub const MAX: usize = usize::MAX;
 /// Subset of `awint::awi`
 pub mod awi {
     // everything except for `char`, `str`, `f32`, and `f64`
-    pub use core::primitive::{
-        bool, i128, i16, i32, i64, i8, isize, u128, u16, u32, u64, u8, usize,
+    pub use core::{
+        assert, assert_eq, assert_ne,
+        primitive::{bool, i128, i16, i32, i64, i8, isize, u128, u16, u32, u64, u8, usize},
     };
 
     pub use Option::{self, None, Some};
     pub use Result::{self, Err, Ok};
+
+    pub use crate::bw;
 }
 
 /// Utility free function for converting a `usize` to a `NonZeroUsize`. This is