+++ testcrate/build.rs
@@ -15,7 +15,7 @@ use std::{
     path::PathBuf,
 };
 
-use awint::prelude::*;
+use awint::awi::*;
 use rand_xoshiro::{
     rand_core::{RngCore, SeedableRng},
     Xoshiro128StarStar,