+++ testcrate/tests/dag_fuzzing.rs
@@ -5,7 +5,7 @@ use awint::{
     awint_dag::{lowering::OpDag, state::STATE_ARENA, EvalError, Lineage, Op, StateEpoch},
     awint_internals::BITS,
     awint_macro_internals::triple_arena::{ptr_struct, Arena},
-    dag, dag_prelude,
+    dag,
 };
 use rand_xoshiro::{
     rand_core::{RngCore, SeedableRng},
@@ -813,58 +813,4 @@ fn dag_fuzzing() {
     let (mut op_dag, _) = OpDag::new(&leaves, &leaves);
     op_dag.render_to_svg_file(std::path::PathBuf::from("rendered.svg"))
         .unwrap();*/
-}
-
-#[test]
-fn state_epochs() {
-    use awint::dag::u8;
-    let state = {
-        let _epoch0 = StateEpoch::new();
-        let x: &u8 = &7.into();
-        // test `Copy` trait
-        let y: u8 = *x;
-        assert_eq!(STATE_ARENA.with(|f| f.borrow().len()), 1);
-        {
-            let _epoch1 = StateEpoch::new();
-            let mut _z: u8 = 7.into();
-            assert_eq!(STATE_ARENA.with(|f| f.borrow().len()), 2);
-        }
-        assert_eq!(STATE_ARENA.with(|f| f.borrow().len()), 1);
-        {
-            let _epoch2 = StateEpoch::new();
-            let mut _w: u8 = 7.into();
-            assert_eq!(STATE_ARENA.with(|f| f.borrow().len()), 2);
-        }
-        assert_eq!(STATE_ARENA.with(|f| f.borrow().len()), 1);
-        let state = y.state();
-        assert!(state.get_state().is_some());
-        state
-    };
-    assert!(state.get_state().is_none());
-    assert!(STATE_ARENA.with(|f| f.borrow().is_empty()))
-}
-
-#[test]
-#[should_panic]
-fn state_epoch_fail() {
-    let epoch0 = StateEpoch::new();
-    let epoch1 = StateEpoch::new();
-    drop(epoch0);
-    drop(epoch1);
-}
-
-#[test]
-fn dag_assertions() {
-    use awint::awint_dag::{assert, assert_eq, assert_ne};
-    use dag_prelude::*;
-    let epoch0 = StateEpoch::new();
-    let x = inlawi!(13u8);
-    let y = inlawi!(13u8);
-    let z = inlawi!(99u8);
-    let is_true = x.lsb();
-    assert!(true);
-    assert!(is_true);
-    assert_eq!(x, y);
-    assert_ne!(x, z);
-    core::assert_eq!(epoch0.assertions().bits.len(), 4);
 }